factor out SplitMaker component
